+++ gradle.properties
@@ -3,6 +3,6 @@ org.gradle.jvmargs=-Xmx1G
 ksp.incremental=false
 
 # Mod Properties
-version=1.0.0-beta.116
+version=1.0.0-beta.117
 group=tech.thatgravyboat
 archives_base_name=skyblock-api
