+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/data/stored/ForgeStorage.kt
@@ -1,7 +1,7 @@
 package tech.thatgravyboat.skyblockapi.api.data.stored
 
 import tech.thatgravyboat.skyblockapi.api.data.StoredProfileData
-import tech.thatgravyboat.skyblockapi.api.profile.mining.forge.ForgeSlot
+import tech.thatgravyboat.skyblockapi.api.profile.items.forge.ForgeSlot
 import tech.thatgravyboat.skyblockapi.generated.SkyblockAPICodecs
 import tech.thatgravyboat.skyblockapi.utils.codecs.CodecUtils
 import tech.thatgravyboat.skyblockapi.utils.codecs.IncludedCodecs
