+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/data/stored/PlayerStorageStorage.kt
@@ -1,9 +1,9 @@
 package tech.thatgravyboat.skyblockapi.api.data.stored
 
 import tech.thatgravyboat.skyblockapi.api.data.StoredProfileData
-import tech.thatgravyboat.skyblockapi.api.profile.storage.PlayerStorageData
-import tech.thatgravyboat.skyblockapi.api.profile.storage.PlayerStorageInstance
-import tech.thatgravyboat.skyblockapi.api.profile.storage.StorageData
+import tech.thatgravyboat.skyblockapi.api.profile.items.storage.PlayerStorageData
+import tech.thatgravyboat.skyblockapi.api.profile.items.storage.PlayerStorageInstance
+import tech.thatgravyboat.skyblockapi.api.profile.items.storage.StorageData
 
 internal object PlayerStorageStorage {
