+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/equipment/EquipmentData.kt
@@ -1,15 +1,10 @@
 package tech.thatgravyboat.skyblockapi.api.profile.equipment
 
-import me.owdding.ktcodecs.GenerateCodec
 import net.minecraft.world.item.ItemStack
-import tech.thatgravyboat.skyblockapi.generated.KCodec
+import tech.thatgravyboat.skyblockapi.RemoveNextVersion
 
-@GenerateCodec
+@RemoveNextVersion
 data class EquipmentData(
     val slots: MutableMap<EquipmentSlot, ItemStack> = mutableMapOf(),
-    val riftSlots: MutableMap<EquipmentSlot, ItemStack> = mutableMapOf()
-) {
-    companion object {
-        val CODEC = KCodec.getCodec<EquipmentData>()
-    }
-}
+    val riftSlots: MutableMap<EquipmentSlot, ItemStack> = mutableMapOf(),
+)
