+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/items/accessory/AccessoryBagData.kt
@@ -1,9 +1,9 @@
-package tech.thatgravyboat.skyblockapi.api.profile.accessory
+package tech.thatgravyboat.skyblockapi.api.profile.items.accessory
 
+import me.owdding.ktcodecs.GenerateCodec
 import net.minecraft.world.item.ItemStack
-import tech.thatgravyboat.skyblockapi.RemoveNextVersion
 
-@RemoveNextVersion
+@GenerateCodec
 data class AccessoryBagItem(
     val item: ItemStack,
     val page: Int,
