+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/items/equipment/EquipmentSlot.kt
@@ -1,11 +1,8 @@
-package tech.thatgravyboat.skyblockapi.api.profile.equipment
+package tech.thatgravyboat.skyblockapi.api.profile.items.equipment
 
-import tech.thatgravyboat.skyblockapi.RemoveNextVersion
 import tech.thatgravyboat.skyblockapi.api.data.SkyBlockCategory
 import tech.thatgravyboat.skyblockapi.utils.extentions.toFormattedName
-import tech.thatgravyboat.skyblockapi.api.profile.items.equipment.EquipmentSlot as NewEquipmentSlot
 
-@RemoveNextVersion
 enum class EquipmentSlot(
     internal val slot: Int,
     vararg categories: SkyBlockCategory,
@@ -28,22 +25,6 @@ enum class EquipmentSlot(
         SkyBlockCategory.BRACELET, SkyBlockCategory.DUNGEON_BRACELET,
     ),
     ;
-
-    companion object {
-        fun fromNewEquipmentSlot(slot: NewEquipmentSlot): EquipmentSlot = when (slot) {
-            NewEquipmentSlot.NECKLACE -> NECKLACE
-            NewEquipmentSlot.CLOAK -> CLOAK
-            NewEquipmentSlot.BELT -> BELT
-            NewEquipmentSlot.GLOVES -> GLOVES
-        }
-    }
-
-    fun toNewEquipmentSlot(): NewEquipmentSlot = when (this) {
-        NECKLACE -> NewEquipmentSlot.NECKLACE
-        CLOAK -> NewEquipmentSlot.CLOAK
-        BELT -> NewEquipmentSlot.BELT
-        GLOVES -> NewEquipmentSlot.GLOVES
-    }
 
     internal val categories: Set<SkyBlockCategory> = categories.toSet()
