+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/items/sacks/SacksData.kt
@@ -1,0 +1,13 @@
+package tech.thatgravyboat.skyblockapi.api.profile.items.sacks
+
+import me.owdding.ktcodecs.GenerateCodec
+import tech.thatgravyboat.skyblockapi.generated.KCodec
+
+@GenerateCodec
+data class SacksData(
+    var items: MutableMap<String, Int> = mutableMapOf(),
+) {
+    companion object {
+        val CODEC = KCodec.getCodec<SacksData>()
+    }
+}
