+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/items/storage/StorageData.kt
@@ -1,0 +1,27 @@
+package tech.thatgravyboat.skyblockapi.api.profile.items.storage
+
+import me.owdding.ktcodecs.GenerateCodec
+import net.minecraft.world.item.ItemStack
+import tech.thatgravyboat.skyblockapi.generated.KCodec
+
+@GenerateCodec
+data class StorageData(
+    val normal: PlayerStorageData = PlayerStorageData(),
+    val rift: MutableList<PlayerStorageInstance> = mutableListOf(),
+) {
+    companion object {
+        val CODEC = KCodec.getCodec<StorageData>()
+    }
+}
+
+@GenerateCodec
+data class PlayerStorageData(
+    val enderchests: MutableList<PlayerStorageInstance> = mutableListOf(),
+    val backpacks: MutableList<PlayerStorageInstance> = mutableListOf(),
+)
+
+@GenerateCodec
+data class PlayerStorageInstance(
+    val index: Int = 0,
+    val items: MutableList<ItemStack> = mutableListOf(),
+)
