+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/mining/forge/ForgeAPI.kt
@@ -1,38 +1,9 @@
 package tech.thatgravyboat.skyblockapi.api.profile.mining.forge
 
-import kotlinx.datetime.Clock
-import me.owdding.ktmodules.Module
-import tech.thatgravyboat.skyblockapi.api.data.stored.ForgeStorage
-import tech.thatgravyboat.skyblockapi.api.events.base.Subscription
-import tech.thatgravyboat.skyblockapi.api.events.base.predicates.InventoryTitle
-import tech.thatgravyboat.skyblockapi.api.events.screen.InventoryChangeEvent
-import tech.thatgravyboat.skyblockapi.utils.extentions.getRawLore
-import tech.thatgravyboat.skyblockapi.utils.extentions.getSkyBlockId
-import tech.thatgravyboat.skyblockapi.utils.extentions.parseDuration
-import tech.thatgravyboat.skyblockapi.utils.regex.RegexGroup
-import tech.thatgravyboat.skyblockapi.utils.regex.RegexUtils.anyMatch
+import tech.thatgravyboat.skyblockapi.RemoveNextVersion
+import tech.thatgravyboat.skyblockapi.api.profile.items.forge.ForgeAPI as NewForgeApi
 
-@Module
+@RemoveNextVersion
 object ForgeAPI {
-
-    private val durationRegex = RegexGroup.INVENTORY.group("forge").create("duration", "Time Remaining: (?<duration>.*)")
-
-    @Subscription
-    @InventoryTitle("The Forge")
-    fun onInvUpdate(event: InventoryChangeEvent) {
-        if (!event.isInMainPart) return
-        if (event.isSkyBlockFiller) return
-
-        val index = event.slot.index - 9
-        val id = event.item.getSkyBlockId() ?: return
-
-        durationRegex.anyMatch(event.item.getRawLore(), "duration") { (duration) ->
-            val duration = duration.parseDuration() ?: return@anyMatch
-            val expire = Clock.System.now() + duration
-            ForgeStorage.setSlot(index, ForgeSlot(id, expire))
-        }
-    }
-
-    fun getForgeSlots(): Map<Int, ForgeSlot> = ForgeStorage.data.toMap()
-
+    fun getForgeSlots(): Map<Int, ForgeSlot> = NewForgeApi.getForgeSlots().map { (key, data) -> key to ForgeSlot.fromNewData(data) }.toMap()
 }
