+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/mining/forge/ForgeSlot.kt
@@ -1,10 +1,18 @@
 package tech.thatgravyboat.skyblockapi.api.profile.mining.forge
 
 import kotlinx.datetime.Instant
-import me.owdding.ktcodecs.GenerateCodec
+import tech.thatgravyboat.skyblockapi.RemoveNextVersion
+import tech.thatgravyboat.skyblockapi.api.profile.items.forge.ForgeSlot as NewForgeSlot
 
-@GenerateCodec
+@RemoveNextVersion
 data class ForgeSlot(
     val id: String,
     val expiryTime: Instant,
-)
+) {
+    companion object {
+        fun fromNewData(data: NewForgeSlot): ForgeSlot = ForgeSlot(
+            id = data.id,
+            expiryTime = data.expiryTime,
+        )
+    }
+}
