+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/storage/StorageData.kt
@@ -1,27 +1,52 @@
 package tech.thatgravyboat.skyblockapi.api.profile.storage
 
-import me.owdding.ktcodecs.GenerateCodec
 import net.minecraft.world.item.ItemStack
-import tech.thatgravyboat.skyblockapi.generated.KCodec
+import tech.thatgravyboat.skyblockapi.RemoveNextVersion
+import tech.thatgravyboat.skyblockapi.api.profile.items.storage.PlayerStorageData as NewPlayerStorageData
+import tech.thatgravyboat.skyblockapi.api.profile.items.storage.PlayerStorageInstance as NewPlayerStorageInstance
+import tech.thatgravyboat.skyblockapi.api.profile.items.storage.StorageData as NewStorageData
 
-@GenerateCodec
+@RemoveNextVersion
 data class StorageData(
     val normal: PlayerStorageData = PlayerStorageData(),
-    val rift: MutableList<PlayerStorageInstance> = mutableListOf()
+    val rift: MutableList<PlayerStorageInstance> = mutableListOf(),
 ) {
     companion object {
-        val CODEC = KCodec.getCodec<StorageData>()
+        internal fun fromNewData(data: NewStorageData): StorageData {
+            return StorageData(
+                normal = PlayerStorageData.fromNewData(data.normal),
+                rift = data.rift.map(PlayerStorageInstance::fromNewData).toMutableList(),
+            )
+        }
     }
 }
 
-@GenerateCodec
+@RemoveNextVersion
 data class PlayerStorageData(
     val enderchests: MutableList<PlayerStorageInstance> = mutableListOf(),
-    val backpacks: MutableList<PlayerStorageInstance> = mutableListOf()
-)
+    val backpacks: MutableList<PlayerStorageInstance> = mutableListOf(),
+) {
+    companion object {
+        internal fun fromNewData(data: NewPlayerStorageData): PlayerStorageData {
+            return PlayerStorageData(
+                enderchests = data.enderchests.map(PlayerStorageInstance::fromNewData).toMutableList(),
+                backpacks = data.backpacks.map(PlayerStorageInstance::fromNewData).toMutableList(),
+            )
+        }
+    }
+}
 
-@GenerateCodec
+@RemoveNextVersion
 data class PlayerStorageInstance(
     val index: Int = 0,
-    val items: MutableList<ItemStack> = mutableListOf()
-)
+    val items: MutableList<ItemStack> = mutableListOf(),
+) {
+    companion object {
+        internal fun fromNewData(data: NewPlayerStorageInstance): PlayerStorageInstance {
+            return PlayerStorageInstance(
+                index = data.index,
+                items = data.items,
+            )
+        }
+    }
+}
