+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/wardrobe/WardrobeData.kt
@@ -1,15 +1,18 @@
 package tech.thatgravyboat.skyblockapi.api.profile.wardrobe
 
-import me.owdding.ktcodecs.GenerateCodec
-import tech.thatgravyboat.skyblockapi.generated.KCodec
+import tech.thatgravyboat.skyblockapi.RemoveNextVersion
+import tech.thatgravyboat.skyblockapi.api.profile.items.wardrobe.WardrobeData as NewWardrobeData
 
-@GenerateCodec
+@RemoveNextVersion
 data class WardrobeData(
     var currentSlot: Int = -1,
     var slots: MutableList<WardrobeSlot> = mutableListOf(),
 ) {
     companion object {
-        val CODEC = KCodec.getCodec<WardrobeData>()
+        fun fromNewData(data: NewWardrobeData) = WardrobeData(
+            currentSlot = data.currentSlot,
+            slots = data.slots.map { WardrobeSlot.fromNewData(it) }.toMutableList(),
+        )
     }
 }
