+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/profile/wardrobe/WardrobeSlot.kt
@@ -1,11 +1,20 @@
 package tech.thatgravyboat.skyblockapi.api.profile.wardrobe
 
-import me.owdding.ktcodecs.GenerateCodec
 import net.minecraft.world.item.ItemStack
+import tech.thatgravyboat.skyblockapi.RemoveNextVersion
+import tech.thatgravyboat.skyblockapi.api.profile.items.wardrobe.WardrobeSlot as NewWardrobeSlot
 
-@GenerateCodec
+@RemoveNextVersion
 data class WardrobeSlot(
     val id: Int,
     val armor: MutableList<ItemStack>,
     val locked: Boolean,
-)
+) {
+    companion object {
+        fun fromNewData(data: NewWardrobeSlot) = WardrobeSlot(
+            id = data.id,
+            armor = data.armor,
+            locked = false,
+        )
+    }
+}
